requirements.txt
pyyaml
jsonschema
pytest
allure-pytest